Updated routes; Merge SaveFeed and SaveFeedTeam; Extract components
