+++ localization/react-intl/src/app/components/feed/FeedPublish.json
@@ -1,0 +1,57 @@
+[
+  {
+    "id": "saveFeed.licenseDetails",
+    "description": "Link to external page with license details",
+    "defaultMessage": "License details"
+  },
+  {
+    "id": "saveFeed.publishTitle",
+    "description": "Title of the section where the publishing preferences are set",
+    "defaultMessage": "Publish to Marketplace <small>(coming soon)</small>"
+  },
+  {
+    "id": "saveFeed.publishBlurb",
+    "description": "Helper text for the publish feed section",
+    "defaultMessage": "Publish your feed to the marketplace to make it discoverable to third-party organizations, while keeping precise control over your assets."
+  },
+  {
+    "id": "saveFeed.publishSwitch",
+    "description": "Label for a switch where the user publishes a feed",
+    "defaultMessage": "Publish shared feed to Marketplace"
+  },
+  {
+    "id": "saveFeed.licenseTitle",
+    "description": "Title of the section where the publishing preferences such as licenses are selected",
+    "defaultMessage": "License"
+  },
+  {
+    "id": "saveFeed.selectLicense",
+    "description": "Error message that appears when a user has tried to submit a form without a legal (copyright) license chosen for their data.",
+    "defaultMessage": "Select a license in order to create and publish this shared feed."
+  },
+  {
+    "id": "saveFeed.learnMoreLicenses",
+    "description": "Link to an external page with more information about the data licenses",
+    "defaultMessage": "Learn more about licenses."
+  },
+  {
+    "id": "saveFeed.licenseBlurb",
+    "description": "Helper text for the license section",
+    "defaultMessage": "A license tells others what they can and can't do with your code."
+  },
+  {
+    "id": "saveFeed.licenseAcademic",
+    "description": "Label for the academic licensing of shared feed data",
+    "defaultMessage": "Academic"
+  },
+  {
+    "id": "saveFeed.licenseCommercial",
+    "description": "Label for the academic licensing of shared feed data",
+    "defaultMessage": "Commercial"
+  },
+  {
+    "id": "saveFeed.licenseOpenSource",
+    "description": "Label for the academic licensing of shared feed data",
+    "defaultMessage": "Open source"
+  }
+]
