+++ localization/react-intl/src/app/components/feed/SaveFeed.json
@@ -1,18 +1,8 @@
 [
   {
-    "id": "saveFeed.licenseDetails",
-    "description": "Link to external page with license details",
-    "defaultMessage": "License details"
-  },
-  {
-    "id": "saveFeed.viewSharedFeed",
-    "description": "Label of a button displayed on the edit feed page that when clicked takes the user to the shared feed page.",
-    "defaultMessage": "View Shared Feed"
-  },
-  {
-    "id": "saveFeed.sharedFeedPageTitle",
-    "description": "Title of the shared feed creation page",
-    "defaultMessage": "Shared feed"
+    "id": "saveFeed.sharedFeedPageSubtitle",
+    "description": "Subtitle of the shared feed creation page",
+    "defaultMessage": "Create a new shared feed"
   },
   {
     "id": "saveFeed.sharedFeedPageEditSubtitle",
@@ -20,9 +10,24 @@
     "defaultMessage": "Edit shared feed"
   },
   {
-    "id": "saveFeed.sharedFeedPageSubtitle",
-    "description": "Subtitle of the shared feed creation page",
-    "defaultMessage": "Create a new shared feed"
+    "id": "saveFeed.sharedFeedPageEditCollabSubtitle",
+    "description": "Subtitle of the shared feed editing page",
+    "defaultMessage": "Collab with likeminded organizations"
+  },
+  {
+    "id": "saveFeed.viewSharedFeed",
+    "description": "Label of a button displayed on the edit feed page that when clicked takes the user to the shared feed page.",
+    "defaultMessage": "View Shared Feed"
+  },
+  {
+    "id": "saveFeed.feedCollaboratorWarning",
+    "description": "Warning displayed on edit feed page when logged in as a collaborating org.",
+    "defaultMessage": "To request changes to this shared feed, please contact the creating organization: {organizer}</strong>"
+  },
+  {
+    "id": "saveFeed.sharedFeedPageTitle",
+    "description": "Title of the shared feed creation page",
+    "defaultMessage": "Shared feed"
   },
   {
     "id": "createFeed.sharedFeedPageDescription",
@@ -58,81 +63,6 @@
     "id": "saveFeed.descriptionLabel",
     "description": "Label for a field where the user inputs text for a description to a shared feed",
     "defaultMessage": "Description"
-  },
-  {
-    "id": "saveFeed.feedContentTitle",
-    "description": "Title of section where a list can be selected as the content of the feed",
-    "defaultMessage": "Feed content"
-  },
-  {
-    "id": "saveFeed.feedContentBlurb",
-    "description": "Helper text for the feed content section",
-    "defaultMessage": "Select a filtered list of fact-checks from your workspace to contribute to this shared feed. You will be able to update this list at any time."
-  },
-  {
-    "id": "saveFeed.feedContentBlurb2",
-    "description": "Helper text for the feed content section",
-    "defaultMessage": "<strong>Note:</strong> Your list must contain <strong>published fact-checks</strong> in order to be part of this shared feed."
-  },
-  {
-    "id": "saveFeed.selectHelper",
-    "description": "Helper text for shared feed list selector",
-    "defaultMessage": "Fact-check title, summary, and URL will be shared with the feed."
-  },
-  {
-    "id": "saveFeed.learnMore",
-    "description": "Link to external page with more details about shared feeds",
-    "defaultMessage": "Learn more."
-  },
-  {
-    "id": "saveFeed.publishTitle",
-    "description": "Title of the section where the publishing preferences are set",
-    "defaultMessage": "Publish to Marketplace <small>(coming soon)</small>"
-  },
-  {
-    "id": "saveFeed.publishBlurb",
-    "description": "Helper text for the publish feed section",
-    "defaultMessage": "Publish your feed to the marketplace to make it discoverable to third-party organizations, while keeping precise control over your assets."
-  },
-  {
-    "id": "saveFeed.publishSwitch",
-    "description": "Label for a switch where the user publishes a feed",
-    "defaultMessage": "Publish shared feed to Marketplace"
-  },
-  {
-    "id": "saveFeed.licenseTitle",
-    "description": "Title of the section where the publishing preferences such as licenses are selected",
-    "defaultMessage": "License"
-  },
-  {
-    "id": "saveFeed.selectLicense",
-    "description": "Error message that appears when a user has tried to submit a form without a legal (copyright) license chosen for their data.",
-    "defaultMessage": "Select a license in order to create and publish this shared feed."
-  },
-  {
-    "id": "saveFeed.learnMoreLicenses",
-    "description": "Link to an external page with more information about the data licenses",
-    "defaultMessage": "Learn more about licenses."
-  },
-  {
-    "id": "saveFeed.licenseBlurb",
-    "description": "Helper text for the license section",
-    "defaultMessage": "A license tells others what they can and can't do with your code."
-  },
-  {
-    "id": "saveFeed.licenseAcademic",
-    "description": "Label for the academic licensing of shared feed data",
-    "defaultMessage": "Academic"
-  },
-  {
-    "id": "saveFeed.licenseCommercial",
-    "description": "Label for the academic licensing of shared feed data",
-    "defaultMessage": "Commercial"
-  },
-  {
-    "id": "saveFeed.licenseOpenSource",
-    "description": "Label for the academic licensing of shared feed data",
-    "defaultMessage": "Open source"
   },
   {
     "id": "saveFeed.updateSaveButton",
