+++ src/app/components/Root.js
@@ -24,7 +24,6 @@ import Published from './team/Published';
 import Spam from './team/Spam';
 import Trash from './team/Trash';
 import SaveFeed from './feed/SaveFeed';
-import EditFeed from './feed/EditFeed';
 import EditFeedTeam from './feed/EditFeedTeam';
 import Feed from './feed/Feed';
 import FeedItem from './feed/FeedItem';
@@ -110,8 +109,7 @@ class Root extends Component {
                   <Route path=":team/suggested-matches(/:query)" component={SuggestedMatches} />
                   <Route path=":team/unmatched-media(/:query)" component={UnmatchedMedia} />
                   <Route path=":team/published(/:query)" component={Published} />
-                  <Route path=":team/feed/:feedId/edit" component={EditFeed} />
-                  <Route path=":team/feed-team/:feedTeamId" component={EditFeedTeam} />
+                  <Route path=":team/feed/:feedId/edit" component={EditFeedTeam} />
                   <Route path=":team/feed/:feedId/:tab(/:query)" component={Feed} />
                   <Route path=":team/feed/create" component={SaveFeed} />
                   <Route path=":team/spam(/:query)" component={Spam} />
