+++ src/app/components/feed/SaveFeed.js
@@ -4,61 +4,24 @@ import { browserHistory } from 'react-router';
 import { createFragmentContainer, graphql, commitMutation } from 'react-relay/compat';
 import Relay from 'react-relay/classic';
 import { FormattedMessage, FormattedHTMLMessage } from 'react-intl';
-import Checkbox from '@material-ui/core/Checkbox';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import styles from './SaveFeed.module.css';
 import FeedCollaboration from './FeedCollaboration';
-import SelectListQueryRenderer from './SelectList';
+import FeedContent from './FeedContent';
+import FeedMetadata from './FeedMetadata';
+import FeedPublish from './FeedPublish';
 import GenericUnknownErrorMessage from '../GenericUnknownErrorMessage';
-import ExternalLink from '../ExternalLink';
 import { FlashMessageSetterContext } from '../FlashMessage';
 import ConfirmProceedDialog from '../layout/ConfirmProceedDialog';
 import Can from '../Can';
 import { getErrorMessageForRelayModernProblem } from '../../helpers';
 import Alert from '../cds/alerts-and-prompts/Alert';
-import SwitchComponent from '../cds/inputs/SwitchComponent';
 import ButtonMain from '../cds/buttons-checkboxes-chips/ButtonMain';
 import TextArea from '../cds/inputs/TextArea';
 import TextField from '../cds/inputs/TextField';
 import TagList from '../cds/menus-lists-dialogs/TagList';
-import SchoolIcon from '../../icons/school.svg';
-import CorporateFareIcon from '../../icons/corporate_fare.svg';
-import OpenSourceIcon from '../../icons/open_source.svg';
 import ChevronDownIcon from '../../icons/chevron_down.svg';
-import FeedMetadata from './FeedMetadata';
-
-const LicenseOption = ({
-  icon,
-  title,
-  description,
-  url,
-  checked,
-  onChange,
-}) => (
-  <div className={styles.licenseOption}>
-    <Checkbox checked={checked} onChange={onChange} />
-    <div className={`${styles.licenseOptionIcon} ${checked ? null : styles.licenseOptionDisabled}`}>
-      {icon}
-    </div>
-    <div>
-      <div className={`typography-subtitle2 ${checked ? null : styles.licenseOptionDisabled}`}>
-        {title}
-      </div>
-      <span className={`typography-body2 ${checked ? styles.licenseOptionDescription : styles.licenseOptionDisabled}`}>
-        {description}
-        {url && (
-          <>
-            &nbsp;
-            <ExternalLink url={url}>
-              <FormattedMessage id="saveFeed.licenseDetails" defaultMessage="License details" description="Link to external page with license details" />
-            </ExternalLink>
-          </>
-        )}
-      </span>
-    </div>
-  </div>
-);
 
 const createMutation = graphql`
   mutation SaveFeedCreateFeedMutation($input: CreateFeedInput!) {
@@ -133,11 +96,25 @@ const destroyMutation = graphql`
   }
 `;
 
+const updateFeedTeamMutation = graphql`
+mutation SaveFeedUpdateFeedTeamMutation($input: UpdateFeedTeamInput!) {
+  updateFeedTeam(input: $input) {
+    feed_team {
+      dbid
+      saved_search_id
+    }
+  }
+}
+`;
+
 const SaveFeed = (props) => {
-  const feed = props.feed || {}; // Editing a feed or creating a new feed
+  const { feedTeam } = props;
+  const feed = feedTeam.feed || {}; // Editing a feed or creating a new feed
+  const isFeedOwner = feedTeam.team_id === feed.team.dbid;
+
   const [title, setTitle] = React.useState(feed.name || '');
   const [description, setDescription] = React.useState(feed.description || '');
-  const [selectedListId, setSelectedListId] = React.useState(feed.saved_search_id);
+  const [selectedListId, setSelectedListId] = React.useState(isFeedOwner ? feed.saved_search_id : feedTeam.saved_search_id);
   const [discoverable, setDiscoverable] = React.useState(Boolean(feed.discoverable));
   const [createdFeedDbid, setCreatedFeedDbid] = React.useState(null);
   const [newInvites, setNewInvites] = React.useState([]);
@@ -245,8 +222,25 @@ const SaveFeed = (props) => {
     });
   };
 
+  const handleSaveFeedTeam = () => {
+    setSaving(true);
+    const input = {
+      id: feedTeam.id,
+      saved_search_id: selectedListId,
+    };
+
+    commitMutation(Relay.Store, {
+      mutation: updateFeedTeamMutation,
+      variables: { input },
+      onCompleted: () => handleViewFeed(feedTeam.feed.dbid),
+      onError: onFailure,
+    });
+  };
+
   const handleConfirmOrSave = () => {
-    if (feed.id || newInvites.length) {
+    if (feed.id && !isFeedOwner) {
+      handleSaveFeedTeam();
+    } else if (feed.id || newInvites.length) {
       setShowConfirmationDialog(true);
     } else {
       handleSave();
@@ -271,6 +265,34 @@ const SaveFeed = (props) => {
     );
   };
 
+  let pageTitle = (
+    <FormattedMessage
+      id="saveFeed.sharedFeedPageSubtitle"
+      defaultMessage="Create a new shared feed"
+      description="Subtitle of the shared feed creation page"
+    />
+  );
+
+  if (feed.dbid) {
+    pageTitle = (
+      <FormattedMessage
+        id="saveFeed.sharedFeedPageEditSubtitle"
+        defaultMessage="Edit shared feed"
+        description="Subtitle of the shared feed editing page"
+      />
+    );
+  }
+
+  if (feed.dbid && !isFeedOwner) {
+    pageTitle = (
+      <FormattedMessage
+        id="saveFeed.sharedFeedPageEditCollabSubtitle"
+        defaultMessage="Collab with likeminded organizations"
+        description="Subtitle of the shared feed editing page"
+      />
+    );
+  }
+
   return (
     <div className={styles.saveFeedContainer}>
       <div className={styles.saveFeedContent}>
@@ -290,6 +312,21 @@ const SaveFeed = (props) => {
               }
             />
           </div> : null }
+
+        { !isFeedOwner && (
+          <Alert
+            variant="warning"
+            title={
+              <FormattedHTMLMessage
+                id="saveFeed.feedCollaboratorWarning"
+                defaultMessage="To request changes to this shared feed, please contact the creating organization: {organizer}</strong>"
+                description="Warning displayed on edit feed page when logged in as a collaborating org."
+                values={{ organizer: feed?.team?.name }}
+              />
+            }
+          />
+        )}
+
         <div>
           <div className={`typography-caption ${styles.sharedFeedTitle}`}>
             <FormattedMessage
@@ -299,19 +336,7 @@ const SaveFeed = (props) => {
             />
           </div>
           <div className="typography-h6">
-            { feed.dbid ? (
-              <FormattedMessage
-                id="saveFeed.sharedFeedPageEditSubtitle"
-                defaultMessage="Edit shared feed"
-                description="Subtitle of the shared feed editing page"
-              />
-            ) : (
-              <FormattedMessage
-                id="saveFeed.sharedFeedPageSubtitle"
-                defaultMessage="Create a new shared feed"
-                description="Subtitle of the shared feed creation page"
-              />
-            )}
+            { pageTitle }
           </div>
           <div className="typography-body1">
             <FormattedMessage
@@ -321,217 +346,89 @@ const SaveFeed = (props) => {
             />
           </div>
         </div>
-        <div className={styles.saveFeedCard}>
-          <div className="typography-subtitle2">
-            <FormattedMessage
-              id="saveFeed.feedDetailsTitle"
-              defaultMessage="Feed details"
-              description="Title of section where the details of the feed are filled. e.g.: title, description"
-            />
-          </div>
-          <FormattedMessage
-            id="saveFeed.titlePlaceholder"
-            defaultMessage="Memorable feed title"
-            description="Placeholder text for feed title field"
-          >
-            { placeholder => (
-              <TextField
-                id="create-feed__title"
-                placeholder={placeholder}
-                label={<FormattedMessage
-                  id="saveFeed.titleLabel"
-                  defaultMessage="Title"
-                  description="Label for the shared feed title input"
-                />}
-                helpContent={<FormattedMessage
-                  id="saveFeed.titleHelper"
-                  defaultMessage="Great shared feed names are short, memorable, and tell your audience the focus of the media"
-                  description="Title input helper text"
-                />}
-                error={noTitle}
-                suppressInitialError
-                value={title}
-                onChange={e => setTitle(e.target.value)}
-                required
-              />
-            )}
-          </FormattedMessage>
-          <FormattedMessage
-            id="saveFeed.descriptionPlaceholder"
-            defaultMessage="Give this shared feed an optional description."
-            description="Placeholder text for feed description field"
-          >
-            { placeholder => (
-              <TextArea
-                id="create-feed__description"
-                placeholder={placeholder}
-                label={<FormattedMessage
-                  id="saveFeed.descriptionLabel"
-                  defaultMessage="Description"
-                  description="Label for a field where the user inputs text for a description to a shared feed"
-                />}
-                value={description}
-                onChange={e => setDescription(e.target.value)}
-              />
-            )}
-          </FormattedMessage>
-          <TagList
-            tags={tags}
-            setTags={setTags}
-          />
-        </div>
-        <div className={styles.saveFeedCard}>
-          <div className="typography-subtitle2">
-            <FormattedMessage
-              id="saveFeed.feedContentTitle"
-              defaultMessage="Feed content"
-              description="Title of section where a list can be selected as the content of the feed"
-            />
-          </div>
-          <div className="typography-body2">
-            <FormattedMessage
-              id="saveFeed.feedContentBlurb"
-              defaultMessage="Select a filtered list of fact-checks from your workspace to contribute to this shared feed. You will be able to update this list at any time."
-              description="Helper text for the feed content section"
-            />
-          </div>
-          <div className="typography-body2">
-            <FormattedHTMLMessage
-              id="saveFeed.feedContentBlurb2"
-              defaultMessage="<strong>Note:</strong> Your list must contain <strong>published fact-checks</strong> in order to be part of this shared feed."
-              description="Helper text for the feed content section"
-            />
-          </div>
-          <SelectListQueryRenderer
-            required={Boolean(feed.id)}
-            value={selectedListId}
-            onChange={e => setSelectedListId(+e.target.value)}
-            onRemove={() => setSelectedListId(null)}
-            helperText={(
-              <span>
-                <FormattedMessage id="saveFeed.selectHelper" defaultMessage="Fact-check title, summary, and URL will be shared with the feed." description="Helper text for shared feed list selector" />
-                &nbsp;
-                <ExternalLink url="https://www.meedan.com">{ /* FIXME update url */}
-                  <FormattedMessage id="saveFeed.learnMore" defaultMessage="Learn more." description="Link to external page with more details about shared feeds" />
-                </ExternalLink>
-              </span>
-            )}
-          />
-        </div>
-        <div className={styles.saveFeedCard}>
-          <div className="typography-subtitle2">
-            <FormattedHTMLMessage
-              id="saveFeed.publishTitle"
-              defaultMessage="Publish to Marketplace <small>(coming soon)</small>"
-              description="Title of the section where the publishing preferences are set"
-            />
-          </div>
-          <span className="typography-body2">
-            <FormattedMessage
-              id="saveFeed.publishBlurb"
-              defaultMessage="Publish your feed to the marketplace to make it discoverable to third-party organizations, while keeping precise control over your assets."
-              description="Helper text for the publish feed section"
-            />
-          </span>
-          <SwitchComponent
-            label={
+
+        { isFeedOwner && (
+          <div className={styles.saveFeedCard}>
+            <div className="typography-subtitle2">
               <FormattedMessage
-                id="saveFeed.publishSwitch"
-                defaultMessage="Publish shared feed to Marketplace"
-                description="Label for a switch where the user publishes a feed"
-              />
-            }
-            checked={discoverable}
-            onChange={() => setDiscoverable(!discoverable)}
-            disabled
-          />
-          { discoverable ?
-            <div className={styles.licenseSection}>
-              <div className="typography-subtitle2">
-                <FormattedMessage
-                  id="saveFeed.licenseTitle"
-                  defaultMessage="License"
-                  description="Title of the section where the publishing preferences such as licenses are selected"
-                />
-              </div>
-              {
-                discoverableNoLicense && (
-                  <Alert
-                    id="save-feed__no-license-error"
-                    title={
-                      <FormattedMessage
-                        id="saveFeed.selectLicense"
-                        defaultMessage="Select a license in order to create and publish this shared feed."
-                        description="Error message that appears when a user has tried to submit a form without a legal (copyright) license chosen for their data."
-                      />
-                    }
-                    content={
-                      <ExternalLink
-                        url="https://www.meedan.com" /* FIXME: Update url */
-                        style={{ color: 'var(--errorSecondary)' }}
-                      >
-                        <FormattedMessage
-                          id="saveFeed.learnMoreLicenses"
-                          defaultMessage="Learn more about licenses."
-                          description="Link to an external page with more information about the data licenses"
-                        />
-                      </ExternalLink>
-                    }
-                    variant="error"
-                  />
-                )
-              }
-              <span className="typography-body2">
-                <FormattedMessage
-                  id="saveFeed.licenseBlurb"
-                  defaultMessage="A license tells others what they can and can't do with your code."
-                  description="Helper text for the license section"
-                />
-                &nbsp;
-                <ExternalLink url="https://www.meedan.com">{ /* FIXME: Update url */}
-                  <FormattedMessage
-                    id="saveFeed.learnMoreLicenses"
-                    defaultMessage="Learn more about licenses."
-                    description="Link to an external page with more information about the data licenses"
-                  />
-                </ExternalLink>
-              </span>
-              <LicenseOption
-                icon={<SchoolIcon />}
-                title={<FormattedMessage
-                  id="saveFeed.licenseAcademic"
-                  defaultMessage="Academic"
-                  description="Label for the academic licensing of shared feed data"
-                />}
-                checked={academicLicense}
-                onChange={() => setAcademicLicense(!academicLicense)}
-                description="Permit the exploration of the data for noncommercial research intended for publication in an academic or other scholarly setting."
-              />
-              <LicenseOption
-                icon={<CorporateFareIcon />}
-                title={<FormattedMessage
-                  id="saveFeed.licenseCommercial"
-                  defaultMessage="Commercial"
-                  description="Label for the academic licensing of shared feed data"
-                />}
-                checked={commercialLicense}
-                onChange={() => setCommercialLicense(!commercialLicense)}
-                description="Permit the use of the data for internal 3rd party business operations, internal research, and development efforts. "
-              />
-              <LicenseOption
-                icon={<OpenSourceIcon />}
-                title={<FormattedMessage
-                  id="saveFeed.licenseOpenSource"
-                  defaultMessage="Open source"
-                  description="Label for the academic licensing of shared feed data"
-                />}
-                checked={openSourceLicense}
-                onChange={() => setOpenSourceLicense(!openSourceLicense)}
-                description="Permits free use and distribution of the data. Enables collaboration and adaptation for various purposes, including commercial uses."
+                id="saveFeed.feedDetailsTitle"
+                defaultMessage="Feed details"
+                description="Title of section where the details of the feed are filled. e.g.: title, description"
               />
             </div>
-            : null }
-        </div>
+            <FormattedMessage
+              id="saveFeed.titlePlaceholder"
+              defaultMessage="Memorable feed title"
+              description="Placeholder text for feed title field"
+            >
+              { placeholder => (
+                <TextField
+                  id="create-feed__title"
+                  placeholder={placeholder}
+                  label={<FormattedMessage
+                    id="saveFeed.titleLabel"
+                    defaultMessage="Title"
+                    description="Label for the shared feed title input"
+                  />}
+                  helpContent={<FormattedMessage
+                    id="saveFeed.titleHelper"
+                    defaultMessage="Great shared feed names are short, memorable, and tell your audience the focus of the media"
+                    description="Title input helper text"
+                  />}
+                  error={noTitle}
+                  suppressInitialError
+                  value={title}
+                  onChange={e => setTitle(e.target.value)}
+                  required
+                />
+              )}
+            </FormattedMessage>
+            <FormattedMessage
+              id="saveFeed.descriptionPlaceholder"
+              defaultMessage="Give this shared feed an optional description."
+              description="Placeholder text for feed description field"
+            >
+              { placeholder => (
+                <TextArea
+                  id="create-feed__description"
+                  placeholder={placeholder}
+                  label={<FormattedMessage
+                    id="saveFeed.descriptionLabel"
+                    defaultMessage="Description"
+                    description="Label for a field where the user inputs text for a description to a shared feed"
+                  />}
+                  value={description}
+                  onChange={e => setDescription(e.target.value)}
+                />
+              )}
+            </FormattedMessage>
+            <TagList
+              tags={tags}
+              setTags={setTags}
+            />
+          </div>
+        )}
+
+        <FeedContent
+          listId={selectedListId}
+          onChange={e => setSelectedListId(+e.target.value)}
+          onRemove={() => setSelectedListId(null)}
+        />
+
+        { isFeedOwner && (
+          <FeedPublish
+            discoverable={discoverable}
+            discoverableNoLicense={discoverableNoLicense}
+            onToggleDiscoverable={() => setDiscoverable(!discoverable)}
+            academicLicense={academicLicense}
+            commercialLicense={commercialLicense}
+            openSourceLicense={openSourceLicense}
+            onToggleAcademic={() => setAcademicLicense(!academicLicense)}
+            onToggleCommercial={() => setCommercialLicense(!commercialLicense)}
+            onToggleOpenSource={() => setOpenSourceLicense(!openSourceLicense)}
+          />
+        )}
+
       </div>
       <div className={styles.saveFeedContentNarrow}>
         <div className={styles.saveFeedButtonContainer}>
@@ -603,7 +500,11 @@ const SaveFeed = (props) => {
 
         <FeedMetadata feed={feed} />
 
-        <FeedCollaboration feed={feed} onChange={setNewInvites} />
+        <FeedCollaboration
+          collaboratorId={feedTeam.team_id}
+          feed={feed}
+          onChange={setNewInvites}
+        />
       </div>
 
       {/* "Delete" dialog */}
@@ -712,7 +613,11 @@ const SaveFeed = (props) => {
                   />
                 </p>
                 <ul>
-                  { newInvites.map(email => <li className={styles.invitedEmail}>&bull; {email}</li>) }
+                  { newInvites.map(email => (
+                    <li key={email} className={styles.invitedEmail}>
+                      &bull; {email}
+                    </li>
+                  ))}
                 </ul>
               </> : null
             }
@@ -732,19 +637,24 @@ const SaveFeed = (props) => {
 };
 
 SaveFeed.defaultProps = {
-  feed: {},
+  feedTeam: {},
 };
 
 SaveFeed.propTypes = {
-  feed: PropTypes.shape({
+  feedTeam: PropTypes.shape({
     id: PropTypes.string,
-    dbid: PropTypes.number,
-    name: PropTypes.string,
-    discoverable: PropTypes.bool,
-    description: PropTypes.string,
-    saved_search_id: PropTypes.number,
-    licenses: PropTypes.arrayOf(PropTypes.number),
-    tags: PropTypes.arrayOf(PropTypes.string),
+    saved_search_id: PropTypes.number.isRequired,
+    team_id: PropTypes.number.isRequired,
+    feed: PropTypes.shape({
+      id: PropTypes.string,
+      dbid: PropTypes.number,
+      name: PropTypes.string,
+      discoverable: PropTypes.bool,
+      description: PropTypes.string,
+      saved_search_id: PropTypes.number,
+      licenses: PropTypes.arrayOf(PropTypes.number),
+      tags: PropTypes.arrayOf(PropTypes.string),
+    }),
   }),
 };
 
@@ -753,26 +663,31 @@ SaveFeed.propTypes = {
 export { SaveFeed };
 
 export default createFragmentContainer(SaveFeed, graphql`
-  fragment SaveFeed_feed on Feed {
+  fragment SaveFeed_feedTeam on FeedTeam {
     id
-    dbid
-    name
-    discoverable
-    description
     saved_search_id
-    licenses
-    tags
-    permissions
-    team {
+    team_id
+    feed {
+      id
       dbid
       name
-      slug
-    }
-    ...FeedCollaboration_feed
-    ...FeedMetadata_feed
-    teams_count
-    saved_search {
-      title
+      discoverable
+      description
+      licenses
+      tags
+      permissions
+      team {
+        dbid
+        name
+        slug
+      }
+      teams_count
+      saved_search_id
+      saved_search {
+        title
+      }
+      ...FeedCollaboration_feed
+      ...FeedMetadata_feed
     }
   }
 `);
